Move files into services & models per the spec
