+++ pkg/model/project.go
@@ -1,5 +1,10 @@
 package model
 
+import (
+	"github.com/goccy/go-yaml"
+	"os"
+)
+
 type ProjectFont struct {
 	Name   string   `yaml:"name"`
 	Styles []string `yaml:"styles"`
@@ -8,4 +13,20 @@ type ProjectFont struct {
 type ProjectFile struct {
 	Name  string        `yaml:"name"`
 	Fonts []ProjectFont `yaml:"fonts"`
+}
+
+func ReadProjectFile(path string) (*ProjectFile, error) {
+	contents, fileErr := os.ReadFile(path)
+
+	if fileErr != nil {
+		return nil, fileErr
+	}
+
+	var projectFile ProjectFile
+
+	if parseErr := yaml.Unmarshal(contents, &projectFile); parseErr != nil {
+		return nil, parseErr
+	}
+
+	return &projectFile, nil
 }
