+++ pkg/service/font.go
@@ -1,4 +1,4 @@
-package font
+package service
 
 import (
 	"fmt"
