+++ pkg/tests/list_test.go
@@ -1,14 +1,14 @@
 package tests
 
 import (
-	"fontman/client/pkg/parser"
+	"fontman/client/pkg/service"
 	"github.com/stretchr/testify/assert"
 	"testing"
 )
 
 func TestSectionParse(t *testing.T) {
 	input := "/test/path/font.ttf: Test Font:style=Regular,Italics"
-	sections := parser.SplitSections(input)
+	sections := service.SplitSections(input)
 
 	if len(sections) != 3 {
 		t.Errorf("Got %d sections, expected 3.", len(sections))
@@ -21,7 +21,7 @@ func TestLineParse(t *testing.T) {
 		/opt/X11/share/fonts/75dpi/helvO14-ISO8859-1.pcf.gz: Helvetica:style=Oblique
 		/opt/X11/share/fonts/75dpi/helvO24-ISO8859-1.pcf.gz: Helvetica:style=Oblique
 		/opt/X11/share/fonts/misc/18x18ko.pcf.gz: Fixed:style=ko`
-	families := parser.ParseListLines(input)
+	families := service.ParseListLines(input)
 	expectedNames := []string{"Ubuntu Mono derivative Powerline", "Courier", "Helvetica", "Helvetica", "Fixed"}
 
 	assert.Equal(t, 5, len(families))
